Load all roles at once in the Admin UI
Load all roles at once in the admin UI so it's  not necessary to scroll
all the way down so the search finds all matching results. Additionally
re-add the angular watchers to the roles list in the series details
because removing them breaks adding custom roles. Also move all changes
to angular-chosen into a directive or delete it completely. Finally,
don't reload roles when switching tabs.