Log the proper index name when updating the asset manager index
When rebuilding the external API indices,
the log is full of messages about updated asset manager items
in the **admin ui** search index.
This is confusing, especially when this takes a long time;
we just thought for a brief moment that we accidentally rebuilt
the admin UI index twice, instead of admin UI and external API,
and because that takes like 3 hours in this case,
that shocked us quite a bit. ;)