Remove OAI-PMH From Default Workflow
This patch removes the OAI-PMH publication from the default workflows to
avoid that users accidentally publish to this channel which by default
allows for unauthorized public access to all media and metadata.

If users want to use OAI-PMH, they need a concept of how to use, harvest
and access this anyway which adding the operation to the workflows forces
them to do. Additionally, usage of this is quite uncommon among
adopters.

This patch intentionally keeps the retract operation to avoid users
retracting something which they previously accidentally published to
OAI-PMH which would leave them with a publication they could not get rid
of. That would be especially problematic for delete operations.