No Public Access Via OAI-PMH
This patch makes the OAI-PMH endpoints private by default to avoid
unauthorized public access to all media and metadata.

This patch is a minimal change for Opencast 7.x and 8.x which just
configures the endpoints to be accessible to `ROLE_ADMIN` only by
default. It should be very easy to adjust if people actually need it.