Update to MariaDB Client
This patch upgrades from the Oracle MySQL JDBC driver ro the MariaDB
JDBC driver to fix licensing issues with the former.

A better long-term solution might be to switch to `pax-jdbc` for
providing the drivers would be easier for additional drivers and even
better for licensing reasons since we are not including the drivers any
longer. But that would need larger changes.