Opencast Plugin Manager
This patch add a new plugin manager, allowing users to easily enable or
disable certain parts of Opencast. This is similar to (and in fact uses)
the Karaf feature configuration with the huge benefit that users to not
have to configure exact versions and keep that updated. Instead they
just need to specify the plugin they want to load.