Remove Outdated Shibboleth Configuration
By now, the Shibboleth configuration is included in the
`mh_default_org.xml` file. Despite that, there still exists an extremely
outdated versions of the security configuration dedicated to Shibboleth
only which is hardly usable and unnecessary due to the main
configuration. This patch removes the outdated configuration.

This is related to  #1116