Asset Upload Title
Configuring custom asset uploads without rebuilding Opencast is nearly
impossible since the title for the upload options is always taken from
the translation files which are build into Opencast.

This minimal patch allows the title to also be specified as part of the
upload configuration, allowing for configuration-only specifications of
new upload options.