Remove compose in favor of encode
By now, the compose workflow operation is doing an exact subset of the
encode workflow operation which universally can be used as a drop-in
replacement.

In addition to the functionality of compose, encode supports parallel
encoding as well as the adaptive encoding profiles but these are
additional features and using the same configuration as compose, encode
works identical. It even uses mostly the same code internally.

So, there is no real reason to keep compose around any longer. That is
why this patch drops it.