Editor should respect the video timeline pointer during zooming
Currently zooming in in the editor will zoom the waveform centered. That means the left boundary will increase same amount of time value as the right one. If the video position pointer is not centered perfectly, as it should not be since opencast 8, you may zoom to a position, where the pointer isn't shown any more. This patch improve this behaviour and respect the video timeline pointer during zooming.