Opencast server node name should be optional
The code is inconsistent about the node name of hosts. The sql scripts (ddl initialization and upgrade) set the property as nullable with varchar length of 255. But in the code we expect an non nullable value with initialized length of 39. Running opencast on a host without node name set (e.g. not configured the property in custom.properties or upgraded from an earlier opencast version) does not enable the servers, systems and services filters in the Admin UI, cause of null values in the oc_host_registration.node_name. This patch fix issues with null values.