403 Logout Option
In case you are logged in but do not have the necessary roles to access
e.g. the admin interface, users will end up on an error page with no
sane way to log out.

This can be pretty problematic if users are unknowingly authenticated
via LTI and then miss the necessary groups.

This patch adds a very simple user interface to convey the necessary
information to users and provide them with an easy way to log out again.