Be more lenient towards service failures
With much more user content like we see right now, it's much more likely
that a failed operation does not mean a broken service or a broken
service configuration.

That is why this pull request increases the number of times a job may
fail on a service before that service is sent into an error state.

Note that services in a warning state will be sanitized automatically
when operations succeed on the same service. Hence, if a service fails
three times but then succeeds, the service will be in a normal state.

The amount of allowed failures chosen is just a guess. A single
operation failing can cause multiple jobs fail simultaneously. That's
why 10 did not seem too high.