Use normal file appender instead of RandomAccessFile for logging
The previous behavior leads to problems with external log rotation
tools. Those tools usually truncate the file to 0 length. When using
this RandomAccessFile, the JVM process keeps a file pointer (a position)
into the file and the next write will write at that position. When
an external process truncates the log file, the next log message is
written at the byte position that's equal to the previous length of
the log file. All previous bytes of the file will then be filled with
0s. This very annoying when trying to open the log file with `less`
or similar tools.

The standard file appender does not maintain such a file pointer and
always correctly appends.

Unfortunately, the normal file appender is apparently slower than
the RandomAccessFile. See these links:

- https://logging.apache.org/log4j/2.x/manual/appenders.html
- https://logging.apache.org/log4j/2.x/performance.html#whichAppender

It looks like at worst, we make logging half as fast (in terms of
messages per second). I am not sure if that will be significant for
Opencast in practice.