Fix LTI Without persistence
This patch fixes LTI which was completely broken if persisting user
references was not enabled and partly broken if it was enabled but the
user was not yet persisted.

This problem was caused by pull request #1532 which this pull request
complete reverts.

In addition, this pull request fixes the problem #1532 was originally
supposed to fix: On concurrent LTI requests it could happen that
Opencast would try to create the same user multiple times in parallel.
This patch fixes the problem by using an optimistic approach,
deliberately allowing an insert to fail, if an entry already exist.