Fix Capture Agent API REST Docs
Pull request #1530 accidentally marked a part of the capture agent aAPI
as deprecated which it certainly is not even though the documentation
was misleading.

The parameter `workflowInstanceId` of the `/ingest/ingest` endpoints is
used by capture agents to identify that an ingest in related to a
specific scheduled event from which later then meta data and attachments
can be retrieved.