Improved ffmpeg profile for extracting the last image of a video in PartialImportWorkflowOperation
The previous ffmpeg profile would lead the PartialImportWorkflowOperation to fail,
if the input video container contains multiple videos.
This can for example occur in webcam recordings from web-conferencing systems, if the browser changes
the webcam resolutions due to i.e. changes in cpu load.

The new profile should be more robust in that regard. It will always extract the last frame
instead of 'by video frame number'. The apparent loss in functionality is not a problem,
as the profile was only used for extracting the last frame anyway.