Window Selection Style
This patch adjust the style of the window size selector for selecting
embed codes in the admin interface.

Before, the selector had a very thick border which looked alien in
comparison to the style commonly used in the admin interface. This patch
switches to thinner, light gray borders which are also used elsewhere in
the ui.