Admin interface event API logs on user error
Requesting an invalid sort criteria from the admin interface like
`/admin-ng/event/events.json?limit=100&offset=0&sort=xy:ASC` causes
Opencast to log long stack traces. This allows users to clutter the
Opencast logs. Since this is a problem with the user request, Opencast
did not do anything wrong.

Therefore, this patch returns the error to the frontend instead and logs
this merely as a debug log.

This fixes  #1751