Fix events sorted by publication
This patch fixes the internal server error when events are sorted by
publication. The error occurred since publication is a complex type
which cannot simply be specified as search criteria.

This patch now uses the amount of publications for sorting.

This is a prime example showing off that 50 different abstraction layers
can easily hinter development. Being able to just specify the required
query would have been so much easier and less obscure.

This fixes #1752