Standalone downloads Paella plugin
On click on the downloads button in the right menu area of the Paella player, all available tracks for download will be shown. The download links are grouped by flavor and ordered by type (audio or video), mimetype and video resolution or audio bitrate in this order. You can also configure tags and/or flavors of tracks, that should be availabe for download. The plugin does not replace the previous downloads tab from description plugin but is an additional standalone plugin. You can enable or disable these both independently.