Clean up basic configuration
This patch removes a bunch of stuff from the basic configuration guide.
The guide was meant to go through the basic configuration steps you
always need and everyone should go through. Advanced configuration does
not belong here and just confuses users.

Furthermore, none of the removed documentation has anything to do with
setting the server URL which is this section's title.