Download button in theodul player (#1719)
Download Tab in Theodul Pplayer

This patch adds the possibility to download videos in the Theodul player. The plugin can be activated in "etc/ui-config/mh_default_org/theodul/config.yml". The configuration name is "allow_video_download". There is also an option that specifies which users are allowed to download a video, it's called "allowed_roles_for_video_download". It contains a regular expression and will only allow users who have at least one role that matches the regex.

Co-authored-by: Martin Wygas <martinwygas@gmai..com>
Co-authored-by: Lars Kiesow <lkiesow@uos.de>