Add Merge Conflict Check
Thsi patch adds a small GitHub Actions workflow which will check if pull
requests have merge conflicts every time another pull request is merged
or a patch is pushed to the Opencast repository.

The action will mark pull request with the `has-conflicts` tag if there
are conflicts and comment on the pull request so that the pull request
author is notified about the problem.