Ensure User Roles
All users accessing Opencast regardless of being logged in or not should
get ROLE_ANONYMOUS to allow for anonymous access to certain public parts
of Opencast. All users logged in to Opencast should have ROLE_USER
regardless of the login method.

This is currently not true for LDAP which instead relies on the
configuration fixing the problem which can have weird side-effects like
a user not being able to watch public content after logging in.

This patch fixes the problem by always providing logged in users with
ROLE_USER and ROLE_ANONYMOUS as users would expect.