PostgreSQL and auto-generated databases
As promised in the proposal “JPA, PostgreSQL and DDL-Scripts” (passed on
Oct 29, 2019) we worked on replacing the current set of MariaDB based
DDL scripts with a proper auto-generated database. These changes are
now merged into the Opencast codebase so that we do not need the DDL
scripts any longer.

This patch brings these changes to an end, removing the DDL scripts and
updating the documentation to reflect the auto-generated database and
the newly added – for now, still marked as experimental – support for
PostgreSQL.

OAI-PMH with it's weird – and somewhat problematic – database usage
still needs a manual set-up of one table. This has been added to the
documentation but is only necessary if you actually configure and use
OAI-PMH which is not the case in a default installation.

To avoid confusion: We still need upgrade scripts. This is only about
the initial database setup.