Remove Dead OpenID Module
This patch removes the opencast-openid module. It seems like it belonged
to Opencast's CAS stack at some point but is not used in any way any
longer. It also does not contain any code but just includes some
dependencies which we then ignore.

This module currently holds the broken dependency which makes Opencast
fail to build.