Drop broken theodul-plugin-timeline-statistics and usertracking
This patch drops the modules theodul-plugin-timeline-statistics and
engage-theodul-plugin-custom-usertracking which are now broken for years and
mostly commented out anyway so that they are still being loaded but provide no
functionality.

If someone wants to fix this at some point, the code is still in our
history. Although, it might also be that people just rewrite this in
somewhat modern JavaScript ;-)