[OpencastToPaellaConverter]: Prefer video
As of now, Opencast/Paella assumes the stream's type to be the same type the last track is.
However, the track-order returned by Opencast appears to depend on various things while processing and
publishing/indexing the video so the order (audio or video last) is hard to determine.

This PR makes Paella Player assume the stream is a video stream as soon as one video track is present.

For reference, see https://groups.google.com/a/opencast.org/g/users/c/e7hP9wANC1w/m/RAwjBBF0AAAJ