Fixing OpenJDK 11 builds by adding missing dependency.
With JDK 9 and newer, the jaxb impl is no longer included by default
in the classpath, hence the complaints in issue 1688.  Adding this
library back to the classpath resolves the issue.  It does not get
picked up correctly by the dependency checker since it is not
directly used, so we add it to the list of manual exclusions.

Fixed #1688