Fix Job Dispatching Test
For a long time parts of the ServiceRegistryJpaImplTest have been
failing occasionally, but on a regular basis:

    [ERROR] Tests run: 11, Failures: 1, Errors: 0, Skipped: 0, Time elapsed: 5.336 s <<< FAILURE! - in org.opencastproject.serviceregistry.impl.ServiceRegistryJpaImplTest
    [ERROR] testDispatchingJobsHigherMaxLoad(org.opencastproject.serviceregistry.impl.ServiceRegistryJpaImplTest)  Time elapsed: 1.45 s  <<< FAILURE!
    java.lang.AssertionError: expected:<http://thirdhost:8080> but was:<null>
	at org.opencastproject.serviceregistry.impl.ServiceRegistryJpaImplTest.testDispatchingJobsHigherMaxLoad(ServiceRegistryJpaImplTest.java:473)

The issue seems to be more likely to pop up on slower machines or if
there is a high load on the machine when running the particular test.

The problem with the assumption that test makes is that it creates a job
and then assumes it is dispatched to a host in a particular amount of
time.  Unfortunately, there is no guarantee for the dispatching to
happen in a given time. This means that from time to time the test will
fail.

One kind of solution would be to slow the test down, making failures
less likely. Still, this does not really fix the underlying false
assumption that it's guaranteed to happen and would also overall slow
down builds which is not nice.

That is why this fix removes the final check instead. That means that
the part where it will test that it jumps out of the job barrier
remains, but checking that it is on a certain host went away.

If that's something we need, we should build a proper test for that
eventually. But for now, not having the test will cause less harm.
That's good.