Limit OpenMP Threads to Prevent Tesseract Blocking the System
This patch is a mitigation for the new Tesseract 4.x locking up the
system during extract text operations.

We have seen systems under full load for up to half a day when running
Tesseract. This seems to be caused by Tesseract spawning far too many
threads. Limiting the OpenMP threads [1] drops the processing time down
to just a few minutes.

Note that this is not completely free of side-effects since it would
also affect other applications run by Opencast which are using OpenMP.
But none of the default ones do and is basically cannot be worse than
what we see with Tesseract right now.

[1] https://www.openmp.org/spec-html/5.0/openmpse58.html