Check for Apereo CLA
This patch adds a bot checking if there exists an individual
contributor license agreement filed with Apereo which has the pull
request author's username attached to it.

If we want this, we should check if the GitHub user names of our regular
contributors are filed correctly since we were on Bitbucket when we
started filing them with Apereo.