WorkflowOperationHandler "CutMarksToSmil"
This operation parses a JSON containing cut marks into a SMIL that can
be used by the VideoEditorWorkflowOperation. It does this by attributing
the given times to the tracks in the given presentation and presenter
flavors. This is useful in case you have cutmarks that need to be
applied to files that are created during a workflow.

Use-case example: Sending raw recording data from a web-conferencing
application to Opencast, where Opencast should first concatenate the
data and then cut the concatenated video file by user-defined cut marks.

Also adds a workflow showcasing the functionality.

Co-authored-by: Lars Kiesow <lkiesow@uos.de>