Add checkstyle rule to enforce consistent bracing style
This follows a proposal by Greg on the mailing list in December 2019.
Almost everyone agreed that we want K&R style braces (no newline before
opening brace) and that we should make braces mandatory. There was no
hard veto.

As with the "120 line limit" rule a couple days ago, this rule is also
disabled for a majority of modules as there are a lot of rule
violations. We expect to correctly format modules one by one and then
remove those style exceptions.