Add checkstyle rule to enforce consistent indentation
We decided on these numbers in the matrix room. The continuation
indentation is also taken from the Eclipse and IDEA style config files.
The `throws` indentation was the most used one across the codebase. If
the `throws` indentation was 4 instead, we would have 745 additional
style violations.

There is one problem though: there is a bug when enforcing the 
continuation indentation when it comes to nested attributes:
https://github.com/checkstyle/checkstyle/issues/4935

I investigated fixing that bug but at least I couldn't do it in a 
reasonable amount of time. So we now decided to not check continuation
indentation strictly. This gets rid of a lot of errors anyway. And
even without that strict check, having more checks is better than not
having checks at all.