Remove dom4j
This patch removes the dom4j dependency which was used solely to provide
an empty XML file for testing but was marked as non-test runtime
dependency. The included version of dem4j has some severe security
vulnerabilities and we should never use this.

Since this only included but never actually used, there is no security
risk.