Shorten Adopter Registration Primary Key
This patch shortens the primary key used for the adopter registration
from `varchar(255)` to `varchar(64)` since some database default
settings seem to cause problems with this and it's used only for short
UUIDs anyway.

If the table was already successfully created with a larger key, this
does not do anything.

This fixes #2282