Limit Incident Text ID Text
This patch shortens the primary key used for the incident text
identifier from varchar(255) to varchar(128) since some database default
settings seem to cause problems with this and it's used only for shorter
strings.

If the table was already successfully created with a larger key, this
does not do anything.

This fixes #2294