Gracefully handle missing Shibboleth User References
Shibboleth will usually check on login if a user is present in the
system, importing the user data as user reference if not or continuing
otherwise.

This behavior fails if a user exists but no user reference, causing the
login process to fail.

This might, for example, be caused if another userprovider is active to
supply the roles which a user has while Shibboleth is used for login.