Auto-generate OAI-PMH database
This addresses my incorrect assumption that OAI-PMH is no longer in any
workflows by default and setting up the database is thus optional. It is
still in the retract workflow which will currently fail if the database
was not manually prepared in advance.

This patch changes the code so that an SQL code is generated which works
with all databases and ensures modifications update the modification
date property on OAI-PMH entries. This will already generate a valid
OAI-PMH repository on its own but is still compatible to the previous
trigger approach.

If a user sets up database trigger, that will ensure that the
modification date is not set in the code but the moment it is actually
being persistent in the database. This works in favor of some less
resilient harvesters. For details see background below.

This does not modify any exising database structure. It will only effect
new database table creation and does not require a migration.

Background: What are the triggers for?
--------------------------------------

OAI-PMH is a harvesting protocol supporting selective harvesting,
meaning that you can e.g. ask only for records being modified after a
given date or time. This allows harvesters to pick up the harvesting
where they last stopped and not harvesting everything every time.

The problem with setting the modification date in code is that an entry
may be written to the database during an active harvest and the
modification date of this new entry may be before the response date of
the harvest. This would mean that if a harvester assumes all records are
contained in the last response, some may get lost.

The Guidelines for Harvester Implementers addresses this problem and
harvesters should deal with this:

- http://www.openarchives.org/OAI/2.0/guidelines-harvester.htm

> Items in a repository may change or be added during a harvest, or
> after a harvest within the same datestamp (i.e. the same day if the
> datestamp is YYYY-MM-DD). This means that to incrementally harvest
> from a repository, a harvester should overlap successive incremental
> harvests by one datestamp increment (i.e. one day if the granularity
> is YYYY-MM-DD). Furthermore, since it is repository implementation
> dependent whether changes that occur during the harvest will be
> reflected in the response, the from argument of the next incremental
> harvest should be based on the the responseDate returned in the first
> partial-list response of a sequence. When harvesting from repositories
> which use a datestamp granularity of one second, it is advisable to
> overlap by a small additional amount to account for any discrepancy
> between the reported responseDate and the time at the repository when
> any search necessary to answer the request was performed.

But at least VideoLaunge did not and the usage of SQL triggers did
protect that harvester implementation from tripping.

This means, the situation after applying this patch is:

- OAI-PMH works out of the box with no modifications
- If you have a very delicate harvester implementation, you can add
  triggers to improve the situation.