OAI-PMH Primary Key Length
This patch fixes the same issue we have had with other database tables
in combination with specific MariaDB configurations of Ubuntu 18.04 and
utf8mb4 encoding. This means that we essentially have to decide between
key length and emoji support. This, obviously, is no choice at all.
Hence, we need to reduce the key size. This is what this patch does.

The key length for `repo_id` can be reduced quite a bit. That works
since the key is hard coded and can have just one value anyway<F3>.

This fixes #2352