Handle image extraction from very short videos
For very short videos (a fraction of a second), the current image
extraction of the studio workflow will fail due to the `fps` filter
dropping all frames.

This patch changes this by going to a higher frame count since this is
just a safe-guard against videos with extremely high frame rates.

This fixes #2360