Perform `check-availibility` WF check with system user
This is to prevent problems when the user starting the workflow is not
authorized to access the distributed assets. This should be rare in
theory, but in practice, the authorization for static files might use
a cache for permissions. This can easily lead to workflows failing
because the the availability check encounters a 403 response due to
permission caching.

Using the system user here doesn't seem like a problem: the check was
never supposed to test any authorization-related things, but simply
whether the published file exists at all.