Correct Series ACLs when Recreating the Search Service Index
When re-creating the search service Solr index (used for engage
publication), series may end up with stricter access control than
expected since series will end up with the access control list of the
last event added to this series.

This may lead to hidden series if some events in this series are hidden,
even if the rest of the events are public.

Note that this bug may only cause stricter ACLs then expected and is
thus no security concern.

This is related to GHSA-vpc2-3wcv-qj4w.