Fix Memory Leak
We noticed that Opencast 9 presenter nodes would leak memory, causing an
eventual Java heap out of memory error and a complete failure of the
node.

The reason behind this seems to be the TrustedHttpClient which tracks
requests and thus needs an explicit closing of each request, else it
will keep those in its internal `responseMap` indefinitely.

This patch does not necessarily fix all occurrences but seems to fix at
least the most prominent ones. A default processing of a single video
caused 39 open (leaked) requests with 37 one in the availability check.