Fixed "No response from service" for videogrid
The videogrid operation would sometimes fail for many videos (>50).
This was due to the WOH sending all the ffmpeg commands to the service as a single job argument, which broke the service if it became to long.
This commit changes the WOH and the service in a way that instead of sending all commands in a single job, each command is now send as its own job.