Make media package handle multi-byte Unicode characters
Parsing media packages had a strange internal behavior of parsing the
XML, possibly modifying the XML, serializing the XML again to then parse
the XML again. The serializer used in this process could not handle
multi-byte Unicode characters properly which broke the entire process.

This was used as part of the search service which caused the Media
module and player to break on such events.

This patch fixes the problem by directly parsing the XML in most cases
and using a different serializer in the other cases.