Series list provider should use admin UI index
The series list provider is designed to be used by the admin UI endpoints.
Admin UI search index is intended to be the data source for list providers.
But the series service is used as a data source instead.
This patch replace the data source of the series list provider by the
admin UI search index.