Autoconfigure Job Dispatching
Our documentation clearly states that job dispatching should be disabled
on all but admin servers to avoid concurrency problems. Nevertheless, we
see problems related to this over and over again because people missed
that.

This patch automatically sets the correct configuration for non-admin
distributions, leading to a safer default configuration.

A code fix removing the job dispatching entirely from non admin modes or
making them work in parallel (e.g. building a failover system) would be
better, but this is an easy solution and should already help a lot.