educast: Don't force inclusion of referred elements in snapshot
When a referred element should be excluded, but the reference has
properties Opencast forces the element to stay in the snapshot. This
changes the behaviour to the same as when no properties are set, i.e.
the refered element will not be included in the snapshot when a user
requests this.

Fixes #2627