Rename `SearchQuery.Sort.DATE_PUBLISHED` to `DATE_MODIFIED`
This more accurately describes what this value is about, especially now
since deleting an element also updates this date.

If we consider the search API a public API (which is de-facto the case),
then this would be a breaking change as the name of the enum variant
is used for the `sort` API parameter. To avoid this breakage, a simple
special case is added to fix up the old name. This fix is temporary
and will be removed in a future version of Opencast.