Reworked Partial Import WOH to use concat demuxer
The concatenation in the Partial Import WOH can require a lot of working memory. Using the concat demuxer instead of concat filter should lower the load considerably.

To that end, all source videos need to be pre-encoded. This commit enforces the use of a preencode encoding profile. It also adds an encoding profile that ensures the necessary prerequisites for concat demuxing.