Fixes removing a series from an event
The frontend attempts to save "no option selected" as `null` to the backend. The backend does not accept `null` values. Instead the frontend should save "no option" as the empty string.

This corrects `null` to the empty string when submitting. A rather crude fix, but I could not figure out why AngularJS fails to set the correct value in this case.