Remove OAI-PMH Harvester
As discussed on list, the OAI-PMH harvester in Opencast was never fully
implemented, is broken in its current state, and has not been worked on
for years. Nevertheless, the code unnecessarily runs in all Opencast
nodes. This patch removes the broken code.

This patch should not effect the OAI-PMH server providing Opencast's
OAI-PMH interface.