Documentation Deployment Conflicts
Since the documentation is deployed via git, concurrent deployment can
cause conflicts and failures. This is not that bad since the system will
automatically heal itself eventually, but reducing the chances for
failures is good.

Most of the time, the problem occurs if multiple patches are merged in
quick succession. In that case, we don't need any but the last one to
succeed anyway, and we can just cancel earlier Deployments of one branch.

This patch does exactly that: Put all workflows run on one branch in a
concurrency group, letting new ones cancel older running ones.