Prevent workflow from failing if we have no logs
When the tests fail before the integration tests are run, the step which
is executed on failures will also fail since it tries to switch into a
non-existing directory.

This is annoying since GitHub Actions will then highlight that error
instead of the one which actually caused the failure.

This patch should fix the problem.