Update Deprecated Code
This is a batch update of old and mostly deprecated code I've updated
since I was working on and with this as part of #2815 anyway.

It shouldn't change how the code works. It mostly fixes logger
statements and replaces custom code with functions from Java which have
been introduced since and which make things a lot easier. This is also
the reason why I code remove a bunch of old utility functions in the
end.