Transformer Factory Identification
This patch updates the identification of the transformer factory used
for XML transformation to use the same method used for the other
implementations. The previous method required the Xalan to be present
which could be annoying problematic for development. This is likely also
the reason why the other libraries use a different mechanism for
identification.