Changed preencode encoding profile to handle divBy2 problems
The encoding profile that is used per default for preencoding in the partial import operation throws an error for videos with a resolution not divisible by 2. As the purpose of preencoding is to avoid exactly this kind of problem, this PR extends the default encoding profile by a crop filter to ensure an even resolution.

This is potentially superseded by #2669.