New Elasticsearch index implementation
This new index implementation replaces the admin UI and External API
indices. It was created in place of AbstractSearchIndex and the module
was renamed accordingly. The old index modules were removed.

Some additional minor changes:

- Some methods were moved from ApiIndex up to AbstractElasticsearchIndex
- getTermsOfField() parameters were simplified
- The EventIndex interface and implementation was removed, instead the
code lives now directly in the list providers as was already the case
for some