Server Job Statistics
This patch reduces the amount of job statistics for servers displayed in
the admin interface to running and queued jobs to avoid performance
problems and remove incorrect and/or misleading data.

Completed jobs were limited to a certain time interval anyway which is
not at all communicated to users and makes the displayed results highly
confusing. In addition to that problem, some completed jobs are removed
from the database, ensuring that the correctness of data even for this
time interval is not guaranteed.