Improve security-jwt tests
This patch improves the stability of some tests in the
`security-jwt` module. There were timing issues in the
`testLoginWithCache` test. Depending on how long the
duration between two statements in this was it failed
or succeeded. This is fixed now. Additionally, I removed
the assertions about exception messages as they might
change quickly. Now only the type of exceptions thrown
is tested.